L3.3

+++ L3/L3.3/internal/models/api_models.go
@@ -4,9 +4,9 @@ import "time"
 
 // CommentNode структура для использования в хендлерах
 type CommentNode struct {
-	ID        int            `json:"id"`
-	UserID    int            `json:"user_id"`
-	ParentID  *int           `json:"parent_id,omitempty"`
+	ID        int64          `json:"id"`
+	UserID    int64          `json:"user_id"`
+	ParentID  *int64         `json:"parent_id,omitempty"`
 	Content   string         `json:"content"`
 	CreatedAt time.Time      `json:"created_at"`
 	Children  []*CommentNode `json:"children,omitempty"`
