+++ L3/L3.3/internal/models/db_models.go
@@ -4,7 +4,7 @@ import "time"
 
 // User модель пользователя
 type User struct {
-	ID        int       `db:"id" json:"id"`
+	ID        int64     `db:"id" json:"id"`
 	Username  string    `db:"username" json:"username"`
 	Email     string    `db:"email" json:"email"`
 	CreatedAt time.Time `db:"created_at" json:"created_at"`
@@ -12,9 +12,9 @@ type User struct {
 
 // Comment модель комментария
 type Comment struct {
-	ID        int        `db:"id" json:"id"`
-	UserID    int        `db:"user_id" json:"user_id"`
-	ParentID  *int       `db:"parent_id" json:"parent_id,omitempty"`
+	ID        int64      `db:"id" json:"id"`
+	UserID    int64      `db:"user_id" json:"user_id"`
+	ParentID  *int64     `db:"parent_id" json:"parent_id,omitempty"`
 	Content   string     `db:"content" json:"content"`
 	CreatedAt time.Time  `db:"created_at" json:"created_at"`
 	UpdatedAt *time.Time `db:"updated_at" json:"updated_at,omitempty"`
